+++ build-logic/gobley-gradle/src/main/kotlin/utils/PluginUtils.kt
@@ -8,14 +8,14 @@ package gobley.gradle.utils
 
 import gobley.gradle.InternalGobleyGradleApi
 import gobley.gradle.PluginIds
+import gobley.gradle.android.GobleyAndroidBaseExtensionDelegate
 import gobley.gradle.android.GobleyAndroidExtensionDelegate
+import gobley.gradle.android.GobleyAndroidKotlinMultiplatformExtensionDelegate
 import gobley.gradle.kotlin.GobleyKotlinAndroidExtensionDelegate
 import gobley.gradle.kotlin.GobleyKotlinExtensionDelegate
 import gobley.gradle.kotlin.GobleyKotlinJvmExtensionDelegate
 import gobley.gradle.kotlin.GobleyKotlinMultiplatformExtensionDelegate
-import gobley.gradle.rust.targets.RustAndroidTarget
 import org.gradle.api.GradleException
-import org.gradle.api.Plugin
 import org.gradle.api.Project
 
 @InternalGobleyGradleApi
@@ -82,10 +82,13 @@ object PluginUtils {
 
     fun withAndroidPlugin(project: Project, action: (GobleyAndroidExtensionDelegate) -> Unit) {
         project.plugins.withId(PluginIds.ANDROID_APPLICATION) {
-            action(GobleyAndroidExtensionDelegate(project))
+            action(GobleyAndroidBaseExtensionDelegate(project))
         }
         project.plugins.withId(PluginIds.ANDROID_LIBRARY) {
-            action(GobleyAndroidExtensionDelegate(project))
+            action(GobleyAndroidBaseExtensionDelegate(project))
+        }
+        project.plugins.withId(PluginIds.ANDROID_KOTLIN_MULTIPLATFORM_LIBRARY) {
+            action(GobleyAndroidKotlinMultiplatformExtensionDelegate(project))
         }
     }
 }
